mcp_server/pyproject.toml
[project]
name = "mcp_server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.9"
dependencies = [
    "mcp[cli]>=1.7.1",
    "psycopg[binary]>=3.2.7",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
]
